+++ NSonarQubeAnalyzer/NSonarQubeDiagnostics/NSonarQubeDiagnostics/Diagnostics/BreakOutsideSwitch.cs
