+++ NSonarQubeAnalyzer/NSonarQubeDiagnostics/NSonarQubeDiagnostics/Diagnostics/EmptyMethod.cs
