+++ NSonarQubeAnalyzer/NSonarQubeDiagnostics/NSonarQubeDiagnostics/Diagnostics/ForLoopCounterChanged.cs
