+++ NSonarQubeAnalyzer/NSonarQubeDiagnostics/NSonarQubeDiagnostics/Diagnostics/IfConditionalAlwaysTrueOrFalse.cs
