+++ NSonarQubeAnalyzer/NSonarQubeDiagnostics/NSonarQubeDiagnostics/Diagnostics/MethodName.cs
