+++ NSonarQubeAnalyzer/NSonarQubeDiagnostics/NSonarQubeDiagnostics/Diagnostics/SwitchWithoutDefault.cs
